Trying to run federated with 2 w
