+++ delicious_lsh_script.sh
@@ -3,12 +3,12 @@
 # Lines that begin with #SBATCH specify commands to be used by SLURM for scheduling
 
 #SBATCH --job-name=fts1.2     # sets the job name if not set from environment
-#SBATCH --time=10:45:00     # how long you think your job will take to complete; format=hh:mm:ss
+#SBATCH --time=20:00:00     # how long you think your job will take to complete; format=hh:mm:ss
 #SBATCH --account=scavenger    # set QOS, this will determine what resources can be requested
 #SBATCH --qos=scavenger    # set QOS, this will determine what resources can be requested
 #SBATCH --partition=scavenger
-#SBATCH --gres=gpu:1
-#SBATCH --ntasks=1
+#SBATCH --gres=gpu:2
+#SBATCH --ntasks=2
 #SBATCH --mem 256gb         # memory required by job; if unit is not specified MB will be assumed
 #SBATCH --nice=0
 #SBATCH --mail-type=END   # Valid type values are NONE, BEGIN, END, FAIL, REQUEUE
@@ -17,4 +17,4 @@ module load mpi
 module load cuda/11.4.4
 module load cudnn/v8.2.1
 
-mpirun -np 1 python run_cluster.py --dataset Delicious200K --cr 0.1 --epochs 5 --batch_size 32 --hash_type slide_avg --sdim 9
+mpirun -np 2 python run_cluster.py --dataset Delicious200K --cr 0.1 --epochs 5 --batch_size 32 --hash_type slide_avg --sdim 9
